+++ EzDeal/API JAVA/src/main/java/com/myorg/ezdeal/repository/HorarioRepository.java
@@ -3,7 +3,7 @@ package com.myorg.ezdeal.repository;
 import com.myorg.ezdeal.models.Horario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HorarioRepository extends JpaRepository<Horario, Integer> {
+public interface HorarioRepository extends JpaRepository<Horario, Long> {
